Add Mocha and a smoke test
I chose the Mocha testing framework since it seems to be one of
the easier ones to get into for a small project.  The smoke test
provided in their Getting Started documentation is included just to
make sure everything is working.  Also included is a script in
package.json for running the tests more easily and some initial
instructions in README.md.